backend/requirements.txt
django
django-ckeditor
pillow
djangorestframework
django-cors-headers